FP8 Scale Location

+++ python/sglang/srt/layers/quantization/marlin_utils_fp4.py
@@ -240,10 +240,10 @@ def prepare_fp4_layer_for_marlin(
     setattr(layer, weight_attr, torch.nn.Parameter(marlin_qweight, requires_grad=False))
 
     # WEIGHT SCALES: Transpose, permute, convert to FP8-S0E5M3 format.
-    # Keep as FP8 until nvfp4_marlin_process_scales to avoid precision loss from
-    # FP8->BF16 conversion before permute (can zero out small scales).
+    # Match vLLM: convert to param_dtype before permute (nvfp4_marlin_process_scales
+    # converts to half first for the comparison, so FP8->param_dtype here is fine).
     weight_scale = getattr(layer, weight_scale_attr)
-    weight_scale = weight_scale.data.T.contiguous()
+    weight_scale = weight_scale.data.T.contiguous().to(param_dtype)
     weight_scale = marlin_permute_scales(
         s=weight_scale,
         size_k=part_size_k,
@@ -345,7 +345,7 @@ def prepare_moe_fp4_layer_for_marlin(layer: torch.nn.Module) -> None:
         scales = getattr(layer, name + "_weight_scale")
         global_scale = getattr(layer, name + "_weight_scale_2")
 
-        # Keep scales as FP8 until nvfp4_marlin_process_scales to avoid precision loss.
+        scales = scales.to(param_dtype)
         global_scale = global_scale.to(param_dtype)
 
         if "w13" in name:
@@ -358,7 +358,6 @@ def prepare_moe_fp4_layer_for_marlin(layer: torch.nn.Module) -> None:
         tensor_list = []
         for i in range(e):
             # Transpose: (size_n, size_k//group_size) -> (size_k//group_size, size_n)
-            # Keep as FP8 until nvfp4_marlin_process_scales to avoid precision loss.
             scale = scales.data[i].T
             marlin_scales = marlin_permute_scales(
                 s=scale,
